+++ example/lib/base/config.dart
@@ -13,7 +13,7 @@ class Config {
 
   static String token = "";              //当前Token
   static String userId = "";             //登录用户ID
-  static String serverAddress = "http://ymbok.com";   //web需要在同一个域名下才能访问接口
+  static String serverAddress = "https://ymbok.com";   //web需要在同一个域名下才能访问接口
 
   static const String LOGIN_PASSWORD_URL =  "/api.php/api/loginPassword/ajax/1";          //登录